Fix the EmptyStringAsConst canonicalizer rule
