+++ src/canonicalizer/rules/empty_string_as_const.h
@@ -1,0 +1,35 @@
+namespace sourcemeta::jsonbinpack::canonicalizer {
+
+class EmptyStringAsConst final : public sourcemeta::alterschema::Rule {
+public:
+  EmptyStringAsConst() : Rule("empty_string_as_const"){};
+  [[nodiscard]] auto
+  condition(const sourcemeta::jsontoolkit::Value &schema,
+            const std::string &dialect,
+            const std::unordered_map<std::string, bool> &vocabularies) const
+      -> bool override {
+    return dialect == "https://json-schema.org/draft/2020-12/schema" &&
+           vocabularies.contains(
+               "https://json-schema.org/draft/2020-12/vocab/validation") &&
+           sourcemeta::jsontoolkit::is_object(schema) &&
+           sourcemeta::jsontoolkit::defines(schema, "type") &&
+           sourcemeta::jsontoolkit::is_string(
+               sourcemeta::jsontoolkit::at(schema, "type")) &&
+           sourcemeta::jsontoolkit::to_string(
+               sourcemeta::jsontoolkit::at(schema, "type")) == "string" &&
+           sourcemeta::jsontoolkit::defines(schema, "maxLength") &&
+           sourcemeta::jsontoolkit::is_integer(
+               sourcemeta::jsontoolkit::at(schema, "maxLength")) &&
+           sourcemeta::jsontoolkit::to_integer(
+               sourcemeta::jsontoolkit::at(schema, "maxLength")) == 0;
+  }
+
+  auto transform(sourcemeta::jsontoolkit::JSON &document,
+                 sourcemeta::jsontoolkit::Value &value) const -> void override {
+    sourcemeta::jsontoolkit::assign(document, value, "const",
+                                    sourcemeta::jsontoolkit::from(""));
+    sourcemeta::jsontoolkit::erase(value, "maxLength");
+  }
+};
+
+} // namespace sourcemeta::jsonbinpack::canonicalizer
